서버

+++ dynamic_web01/src/main/java/controller/BookmarkDeleteController.java
@@ -7,7 +7,9 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import model.dto.BookmarkDTO;
+import model.dto.UserDTO;
 import model.service.BookmarkService;
 
 @WebServlet("/bookmark/delete")
@@ -15,18 +17,28 @@ public class BookmarkDeleteController extends HttpServlet {
 	
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		HttpSession session = req.getSession();
+		
+		if(!(boolean) session.getAttribute("login")) {
+			resp.sendRedirect(req.getContextPath() + "/login");
+			return;
+		}
+		
+		UserDTO userData = (UserDTO)session.getAttribute("user");
+		
 		String id = req.getParameter("id");
 		
 		BookmarkDTO dto = new BookmarkDTO();
 		dto.setId(Integer.parseInt(id));
+		dto.setUserId(userData.getUserId());
 		
 		BookmarkService service = new BookmarkService();
 		boolean result = service.delete(dto);
 		
 		if(result) {
-			resp.sendRedirect("../bookmark");
+			resp.sendRedirect(req.getContextPath() + "/bookmark");
 		} else {
-			resp.sendRedirect("../error");
+			resp.sendRedirect(req.getContextPath() + "/error");
 		}
 	}
 }
